[DHS-1682]: add util function to get project from collection path
Needed for DHS-1709 "Disable 'Unarchive' buttons if this AVU is set to 'True'"

When we generate the data (ajax.py) to display the list of files, we get
sent the path to a collection. But the enableUnarchive exists at the
project level.

It's not pretty as I think this should be centralized. Also there might
already be some facility in place to do this.